Add auto mode

+++ scripts/modules/allskymodulemanager/moduleinstaller.py
@@ -196,7 +196,46 @@ class ALLSKYMODULEINSTALLER:
             
         return modules_to_install
     
-    def cleanup_opt(self):
+    def auto_upgrade_modules(self, args):
+        shared.log(4, f"Auto upgrade modules started")
+        shared.log(4, f"============================\n")
+
+        if args.dryrun:
+            shared.log(4, f"WARNING: Using dry run mode, no changes will be made\n")
+            
+        shared.log(4, f"Initialising Allsky Module repository")
+        shared.log(4, f"=====================================\n")            
+        self._ensure_cloned_repo(self._module_repo, self._module_repo_path, self._branch, True)
+        self._branches = self._get_remote_branches(self._module_repo_path)
+        self._select_git_branch(args)
+
+        shared.log(4, f"\nReading Installed modules")
+        shared.log(4, f"=========================\n")
+        self._read_modules()
+        
+        shared.log(4, f"\nUpdating modules")
+        shared.log(4, f"================\n")
+                
+        for module in self._module_list:
+            try:
+                if module.installed:
+                    shared.log(4, f"INFO: Updating {module.name}")
+                    if not args.dryrun:
+                        module.install_or_update_module(True)
+            except Exception as e:
+                tb = e.__traceback__
+                shared.log(4, f"ERROR: Function auto_upgrade_modules on line {tb.tb_lineno}: {e}")
+        
+        shared.log(4, "INFO: Auto upgrade modules completed\n\n")
+    
+        self.cleanup_opt(args)
+        
+    def cleanup_opt(self, args):
+        
+        did_something = False
+        shared.log(4, f"Starting cleanup of /opt/allsky/modules")
+        shared.log(4, f"=======================================\n")
+        
         try:
             allsky_module_folder = shared.get_environment_variable("ALLSKY_MYFILES_DIR")
             allsky_module_folder = os.path.join(allsky_module_folder, "modules")   
@@ -207,45 +246,57 @@ class ALLSKYMODULEINSTALLER:
             opt_module_dependencies_folder_base = os.path.join(opt_module_folder, "dependencies")
             opt_module_folder = Path(opt_module_folder)
 
-            if not allsky_module_folder.is_dir() or not opt_module_folder.is_dir():
-                return
-
             for opt_module in opt_module_folder.iterdir():
                 if opt_module.is_file():
                     allsky_module = allsky_module_folder / opt_module.name
                     if allsky_module.exists():
                         module_name = os.path.splitext(os.path.basename(opt_module))[0]
                         shared.log(4, f"INFO: Removing {opt_module} and its info and dependencies folders")
-                        os.remove(opt_module)
-                        opt_module_info_folder = os.path.join(opt_module_info_folder_base, module_name)
-                        opt_module_dependencies_folder = os.path.join(opt_module_dependencies_folder_base, module_name)
-                        shared.remove_path(opt_module_info_folder)
-                        shared.remove_path(opt_module_dependencies_folder)
+                        if not args.dryrun:
+                            os.remove(opt_module)
+                            opt_module_info_folder = os.path.join(opt_module_info_folder_base, module_name)
+                            opt_module_dependencies_folder = os.path.join(opt_module_dependencies_folder_base, module_name)
+                            shared.remove_path(opt_module_info_folder)
+                            shared.remove_path(opt_module_dependencies_folder)
+                        did_something = True
 
             
             if os.path.isdir(opt_module_info_folder_base):        
                 if not any(Path(opt_module_info_folder_base).iterdir()):
                     shared.log(4, f"INFO: Removing empty directory {opt_module_info_folder_base}")
-                    shared.remove_path(opt_module_info_folder_base)
+                    if not args.dryrun:
+                        shared.remove_path(opt_module_info_folder_base)
+                    did_something = True
             if os.path.isdir(opt_module_dependencies_folder_base):                   
                 if not any(Path(opt_module_dependencies_folder_base).iterdir()):
                     shared.log(4, f"INFO: Removing empty directory {opt_module_dependencies_folder_base}")                    
-                    shared.remove_path(opt_module_dependencies_folder_base)
+                    if not args.dryrun:
+                        shared.remove_path(opt_module_dependencies_folder_base)
+                    did_something = True
 
             py_cache_path = os.path.join(self._module_repo_base_path, "modules", "__pycache__")
             if os.path.isdir(py_cache_path):
-                shared.remove_path(py_cache_path)
-
+                shared.log(4, f"INFO: Removing python cache directory {py_cache_path}") 
+                if not args.dryrun:
+                    shared.remove_path(py_cache_path)
+                did_something = True
+                
             if opt_module_folder.exists() and opt_module_folder.is_dir():     
                 for file in opt_module_folder.iterdir():
                     if file.is_file() and file.name.startswith("allsky_"):
                         target = allsky_module_folder / file.name
-                        shared.log(4, f"INFO: Moving {file} → {target}")
-                        shutil.move(str(file), str(target))                                        
+                        shared.log(4, f"INFO: Moving non Allsky module {file.name} from {os.path.dirname(file)} to Allsky module folder")
+                        if not args.dryrun:
+                            shutil.move(str(file), str(target))
+                        did_something = True                                    
         except Exception as e:  
             tb = e.__traceback__
             shared.log(4, f"ERROR: Function cleanup_opt on line {tb.tb_lineno}: {e}")
-            
+
+        if not did_something:
+            shared.log(4, "INFO: No cleanup required")
+        shared.log(4, f"INFO: Completed cleanup of /opt/allsky/modules")
+                    
     def run(self, args: argparse.Namespace) -> None:
         self._ensure_cloned_repo(self._module_repo, self._module_repo_path)
         self._branches = self._get_remote_branches(self._module_repo_path)
@@ -262,13 +313,15 @@ class ALLSKYMODULEINSTALLER:
             w = Whiptail(title="Main Menu", backtitle=f"Allsky Module Manager. Using branch {self._branch}", height=20, width=40)
             menu_options = ['Install Modules', 'Uninstall Modules', 'Exit']
             if self._debug_mode:
-                menu_options = ['Install Modules', 'Uninstall Modules', 'Switch Branch', 'Exit']
+                menu_options = ['Install Modules', 'Uninstall Modules', 'Switch Branch', 'Clean Opt', 'Exit']
                 
             menu_option, return_code = w.menu('', menu_options)
 
             if return_code == 0:
                 if menu_option == 'Exit':
                     done = True
+                if menu_option == 'Clean Opt':
+                    self.cleanup_opt()
                 if menu_option == 'Switch Branch':
                     self._select_git_branch(args, True)
                     self._read_modules()
@@ -294,15 +347,27 @@ if __name__ == "__main__":
     parser.add_argument("--branch", action="store_true", help="Allow the remote branch to be selected, default is Master")
     parser.add_argument("--setbranch", type=str, help="Specify the remote branch to use")
     parser.add_argument("--cleanupopt", action="store_true", help="Cleanup the legacy module folders")
+    parser.add_argument("--auto", action="store_true", help="Auto upgrade modules, will migrate if required")    
+    parser.add_argument("--dryrun", action="store_true", help="For auto mode do a dry run only, no changes will be made")    
     args = parser.parse_args()    
     
     module_installer = ALLSKYMODULEINSTALLER(args.debug)
+    
+    if args.auto:
+        try:
+            module_installer.auto_upgrade_modules(args)
+        except Exception as e:
+            tb = e.__traceback__
+            print(f"ERROR: Function auto_upgrade_modules on line {tb.tb_lineno}: {e}")
+        sys.exit(0)
+                
     if args.cleanupopt:
         module_installer.cleanup_opt()
-    else:
-        #try:
+        sys.exit(0)
 
-        module_installer.run(args)
-        #except Exception as e:
-        #    print(e)
+    #try:
+
+    module_installer.run(args)
+    #except Exception as e:
+    #    print(e)
